removes module named "common"
