+++ api/src/main/java/com/flipkart/flux/api/WorkflowStateSummary.java
@@ -11,7 +11,7 @@
  * limitations under the License.
  */
 
-package com.flipkart.flux.commons.dto;
+package com.flipkart.flux.api;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
