+++ pom.xml
@@ -60,7 +60,6 @@
     <modules>
         <module>model</module>
         <module>persistence-mysql</module>
-        <module>commons</module>
         <module>runtime</module>
         <module>task</module>
         <module>client</module>
