+++ runtime/pom.xml
@@ -65,12 +65,6 @@
             <artifactId>task</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
-        <dependency>
-            <groupId>com.flipkart.flux</groupId>
-            <artifactId>commons</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
         <!-- Third party dependencies -->
 
         <dependency>
